Add lead currents to lv2 mutual inductor
The leve 2 nonlinear mutual inductor did not support
lead currents -- this commit adds that functionality.
This issue arose from working on comparison test
cases for the level 1, nonlinear mutual inductor
with other codes.  I'm still try to see if the level 2
inductor can be used as a replacement for the level 1
inducotr.  The level 2 uses an implicit formulation for
the magnetic moment which could be more stable on some
circuits.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/105