Add Support for FIND-AT Measure for .MEASURE FFT
This commit adds support for the FIND-AT measure for .MEASURE FFT.
It is only supported for V, VR, VI, VM, VP, VDB, I, IR, II, IM, IP
and IDB operators.  It is not currently supported if the measured
output quantity is a P or W operator, or an expression.  For this
measure type, the V() and I() operators default to the VM() and
IM() operators, respectively, to match HSPICE.  The requested
AT value is rounded to the value of the nearest harmonic, and the
stdout provides both the requested value and that rounded value
for each FIND-AT measure.

At present, complex operators like VR and II are not supported for
the ENOB, SFDR, SNDR and THD measure types.  Also, expressions and
the P and W operators are not supported for the FIND-AT measure
type.  FIND-WHEN measures are also not supported.

This commit also did a re-factoring of how the FFTAnalysis object
calculates its various metrics.  It introduced two new vectors,
fftRealCoeffs_ and fftImagCoeffs_, which are the un-normalized FFT
coefficients translated into one-sided HSPICE format.  The output
(ftOutData_) of the Xyce FFT interface into converted into those
two vectors, and the remainder of the calculations are then done
on those one-sided FFT coefficient vectors instead.  This makes
those calculations clearer, and simplifies the "getters" needed by
the FIND-AT measure for FFT measure mode.

SON Bugs 1280 and 1327
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1327