Fix non-physical B-H loop output
Fix the calculation of the B-H output for when a
current temporally apposes the magnetic field.  This
results in a correct step turn on the B-H output
rather than a negitive slope in dB/dH as Xyce
as producing.  This was fixed earlier for the
level 1 non-linear mutual inductor and this applies
similar logic to the level 2 non-linear mutual
inductor.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/105
SON bugzilla 991.