Make .STEP Work with .FFT and .MEASURE FFT
This commit enables .FFT and .MEASURE FFT to work with .STEP analyses.
The technical approach used to insert breakpoints at the requested FFT
sample times for .OPTIONS FFT fft_accurate=1, at the first call to the
FFTAnalysis::updateFFTData() function for each step loop, is inelegant
and may be re-factored at some point.  It was chosen because the
StepErrorControlStep object clears its list of breakpoints at the
start of each .STEP loop; and the previous spot for that insertion (in
the FFTAnalysis::fixupFFTParameters()function), is not called at the
start of each .STEP iteration.  Second, the call to Transient::notify()
occurs after the current call to FFTAnalysis:notify().  So, any
re-insertion of those breakpoints in FFTAnalysis:notify() gets
cleared by the call to StepErrorControl::resetAll() by
Transient::notify().

This commit also changes the extension of the .FFT output from
<netlistName>.fft to <netlistName>.fftX where X is the step
number.  This is consistent with how Xyce handles .MEASURE output.
It is NOT consistent with how HSPICE numbers its FFT output.  In
HSPICE, each .FFT line produces its own <netlistName>.fftX file,
where X runs from 0 to the (#FFT lines in netlist)-1.

The Xyce .FFT output should now be usable and correct with these
caveats:

  a) The FREQ, FMIN and FMAX qualifiers will parse, but they have
no effect on the output.

  b) The START and STOP qualifiers will parse, but may produce
different answers then HSPICE does.

  c) The GAUSS and KAISER windows are not implemented yet, and
default to RECT.

  d) The Harmonic List and SNR metrics, for .OPTIONS FFT FFTOUT=1
output, are not produced yet.

SON Bugs 1280 and 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327