Replace deprecated mem_fun with mem_fn
The old standard function std::mem_fun creates a wrapper function to a
classes' member function given a pointer to that function.

Use of this feature was introduced into the MeasureManager by Dave
Baur in commit ac5bbebf back in 2013.

mem_fun is in fact deprecated in C++11 and will be removed in C++17.
It has been replaced by the more functional and more flexible mem_fn.

This commit simply replaces mem_fun usage with mem_fn.

Closes #134
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/134