Change Implementation of BLACK Window Type for .FFT
This commit changes the implementation of the BLACK window type.
It had been using the "Conventional Blackman Window" definition,
which is the "2 digit place approximation" to the Exact Blackman
window.  It now uses the "-67 dB Three-Term Blackman-Harris"
window, as listed in SAND2017-4042.  This provides better HSPICE
compatibility.

SON Bugs 1327 and 1329
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1329