Comments out random assignment of voltage node ownership
In parallel, the assignment of voltage nodes that are shared between
two processors is determined randomly, using a random shuffle.  This
means that any two executions of the simulator, in parallel, can have
slightly different behavior if shared nodes are assigned to different
processors.  While testing should be more resilient to non-deterministic
behaviors in the code and not use diffs on output.  This behavior
should be left to the linear algebra package and the device model
package, not topology.

Removing this behavior addresses some of the regression test failures
that are slowing down parallel nightly testing that uses increased
numbers of processors (4, 8, etc.)

BUG 1239 (SON)