Support FREQ Qualifier and SNR Metric for .FFT Lines
This commit adds support for the FREQ qualifier on .FFT lines, and
for the SNR metric in the optional .OPTIONS FFT FFTOUT=1 output.
The FREQ qualifier allows the user to designate a frequency,
other than the "fundamental frequency" of 1/(stopTime_ - startTime_),
as the "first harmonic".  (Note: FREQ is rounded to the nearest
integer multiple of the fundamental frequency.)   The output to the
<netlistName>.fftX file for that .FFT line then only contains the
magnitude and phase information for the harmonics (e.g., 2,4,6, ...
if FREQ=2) of the frequency specified by the FREQ qualifier.

The calculations for the THD, SFDR, SNDR and ENOB were moved to
their own functions.  A new calculation for the SNR was added. This
is a bit less efficient, since the magnitude vector is now looped
over 5 extra times.  However, it makes it clearer how those five
metrics are calculated if FREQ is given and not equal to the
fundamental frequency.  It also makes it easier to individually
tweak each metric.

The SNR is now output as part of the optional FFTOUT=1 metrics.
If FREQ is equal to the fundamental frequency (which is the default
case if FREQ is not given on the .FFT line), then there are no "noise
frequencies" to include in the SNR calculation.  So, in that case,
the SNR is set using the noiseFloor_ variable (which is currently
set to 1e-10) for compatibility with HSPICE.

SON Bugs 1327 and 1330
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1330