Fixes parallel hanging in packed restarting
When the parallel regression tests are run on more than 4 processors,
the BUG_456 restart tests start hanging.  This is due to a bug in the
assignment of restart nodes to processors.

1. The maximum number of restart nodes allocated to each processor was
incorrectly calculated.

2. If a processor was not allocated any restart nodes, it should be
passed a flag (buffer size of zero) that indicates no restart nodes
are being assigned to it. [This is where the hang was happening]

3. Once a processor is informed it won't be assigned any restart nodes,
it should allocate an integer size buffer and place a zero in that buffer.
Then, when the buffers are sent around the ring so that every processor
can restore their devices, it is a valid buffer.

Gitlab issue #137