Document issue #128
Issue #128 on gitlab-ex concerns an obscure bug in ADMS that resulted
in Xyce/ADMS emitting improper code for very special cases of "for" or
"while" loops.

The issue has never surfaced in any CMC standard model (or other
normal model) but was reported by an open source user who was trying
to compile Verilog-A models buried in a proprietary PDK.

This commit documents that the issue was fixed.

Closes #128
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/128