Removes assembleGlobalVector from block vector classes
Since block vectors are used after assembly, they don't have the
overlap and assembled map issues of the vector and multi-vector
classes.  The assembleGlobalVector method was created to be used
at synchronization points where copying between two different
vectors was necessary.  This is not used, so it is being removed
from the block vector classes.

Gitlab issue #36