Add nint() and fmod() to the expression library
Issue #139.

This was reported by a user, and is needed for HSpice compatibility for
a modern PDK.

Adding nint() and fmod() was pretty easy to do.  nint(x) is a rounding
function, that rounds x to the nearest integer.  fmod(a,b) returns the remainder
of the division of a by b as a real number.

Both of these have corresponding functions in the C++ standard library,
so all I had to do was create 2 new tokens in the lexer
(ExpressionLexer.l), 2 new rules in the parser (ExpressionParser.yxx)
and two new AST classes in the ast.h file.  I have also created a bunch
of unit tests in both the parserUnitTest.C program and the
complexParserUnitTest.C program.

So, other than creating nightly regression tests, this issue is done.