Add Support for -remeasure for .MEASURE FFT
This commit adds support for -remeasure for .MEASURE FFT, and also
for .FFT lines since re-processing the .FFT lines in the netlist is
needed to recalculate any FFT mode measures.  This commit supports
both the step and non-step cases.

Since the StepErrorControl object is not instantiated by the
AnalysisManager during -remeasure, the fft_accurate_ variable
in the various FFTAnalysis objects must be set to 0 during
remeasure.  If the re-measured output file (e.g., <netlistName>.prn)
was previously made with fft_accurate_=1 then the waveform values
at those requested sample times should be in that output file.  So,
in that case, the interpolation done with fft_accurate_ set to 0
should get the same answer as the original netlist.

SON Bugs 1280 and 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327