Adding a header check for Amesos2_Basker.hpp
To improve the robustness of building the linear solvers, with and
without supporting solvers in Trilinos, an additional check is being
added to detect if the templated Basker solver in Amesos2 has been
enabled.  This will allow the removal of the SHYLUBASKER flag that
is confusing and only necessary for Trilinos 12.12.

Gitlab issue #99