Make CMake recognize ShyLU, Amesos2, and Basker(s)
Prior to this change, the searches performed by CMake were not
compatible with the way Xyce was set up to recognize the Amesos2 Basker
solver.  This commit should align everything. It also fixes the search
for the ShyLU Basker solver.

For the record, the templated Basker solver is in Amesos2. If CMake
finds that, it enables the `Xyce_AMESOS2_BASKER` flag. The
multi-threaded Basker solver is part of ShyLU. If CMake finds that, it
enables the `Xyce_SHYLU_BASKER` flag. (The ShyLU Basker is in its own
subpackage, called "ShyLU_NodeBasker".) Right now nothing in Xyce
leverages the `Xyce_SHYLU_BASKER` flag, so it has not been added to the
`Xyce_config.h.cmake` file.

In Trilinos 12.12, one could not build both Baskers at the same time.
The Xyce preprocessor logic used this by enabling Amesos2 Basker when
ShyLU Basker wasn't found. This commit changes that logic to be simply
`#ifdef Xyce_AMESOS2_BASKER`.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/99