Changes the block vector class to an abstract base and Epetra implementation
The Linear::BlockVector class is now mostly abstract.  The Epetra
implementation in the Linear::BlockVector class is now in the
Linear::EpetraBlockVector.  The block system helper functions now
create an EpetraBlockVector object.  A few remnants of Epetra
remain due to the fact that Linear::Vector still is based on
Epetra.

Gitlab issue #36