Use ifdef with Xyce_STOKHOS_ENABLE instead of if
When `Xyce_STOKHOS_ENABLE` was defined, all the preprocessor macros used
```
 #if Xyce_STOKHOS_ENABLE
```
rather than `#ifdef`. This is inconsistent with all the other
preprocessor macros in Xyce, and forced the use of `cmakedefine01`
instead of just `cmakedefine` in the CMake system. This is a purely
cosmetic change, but code consistency is a good thing.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/99