Add .FFT and .OPTIONS FFT to Ref. Guide
This commit adds some initial text on .FFT and .OPTIONS FFT to
the Reference Guide.  The .FFT capability should now be usable,
with the caveats on unsupported features and differences with
HSPICE noted in the .FFT section.  Additional text is still
need on how the ENOB, SFDR, SNR, SNDR and THD metrics are
calculated.  A Users' Guide section should also be added, with
some annotated examples of the <netlistName>.fftX file output.

SON Bug 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327