Over-rides all inherited methods from Vector/MultiVector
All of the methods from the Vector and MultiVector classes are
being implemented in the EpetraBlockVector class.  These methods
use the Epetra_MultiVector object in the MultiVector base class.
The next step is to create a local Epetra_MultiVector object and
change the EpetraBlockVector objects to use it.

Gitlab issue #36