Expanded support for fasttable
This is related to work I did a little while ago, to create a "fasttable"
capability to compliment "table".  The difference is that "fasttable" doesn't
use very many breakpoints; just one of the first and last point in the
table.

I previously set up rules for fasttable to work with tables that are
read in from a file.  But I hadn't properly set it up for other case,
such as when a table is specified directly in the netlist.  This
commit corrects that oversight.