Fix the break point issues in the delay device
The break points were not set up correctly in either the delay
device or the lossless transmission line device. If there is a
large change in the slope of the controlling signal (discontinuous
point/break point), the output of a delay device should insert
a break point at a delayed time. However, the large slope change
is only detected after Xyce takes a step out of a break point. This
time step can be larger than the delayed. In this case,
the break point is ignored and this leads to spikes in the output
signal due to the high order interpolation.

In other cases, this can also lead to wrong answers due to the incorrect
order of when the break points and stop times are updated in the transient
simulation.

To fix the problem, isConverged() function is implemented for the delay
device. If the large slope change is detected and the time step is
larger than the delay, the device is not converged and therefore the
time step is rejected. This way, the first step out of a break point is
less than the delay time. So a break point at a delayed time can be inserted after the
first step out of a discontinuous point in Xyce