Fix the device convergence check in the nonlinear solver
The nonlinear solver checks a list of things to decide if newton
converges. One of criterion is checking if all devices converge.
If they are not converged, then newton continues to the next iteration.
However, the device convergence check only happens when the number
of newton iteration is less than the maximum number of iteration. If
the current iteration number is equal or larger than the maximum, the
device convergence is not checked. Newton can converge incorrectly if
not all devices are converged. This commit fixes that