Improve break point handling in the delay device
The delay device inserts a new breat point at the delayed time when
a discontinuous behavior is detected with the recent
fixes. However, if the first step out of a discontinuou
point is at or very close to the delayed break point, no break
point should be inserted. This commit prevents unnecessary reduction of the
time step and tiny time steps which can cause problems