Adds check for Amesos2-Basker solver class name
In Trilinos 12.12 the Basker solver in the Amesos2 package could not
be compiled in the same binary as the ShyLU-Basker nodal solver because
of a naming conflict.  This was resolved after Trilinos 12.12, but it
has to be detected, if the Amesos_Basker.hpp header is found in the
installed Trilinos.  A new check will see if the old class name is
being used in that header.