Fix the transient analysis to properly handle break points
The transient analysis class is in charge of updating break
points, min and max steps, stop time, calling time integrator
to determine next time steps, etc. The information updates and
calling time integrator need to be in a proper order. The current
order only works for the pre-dertermined break points. Some
devices, such as delay device and transmission lines, can dymamically
add break points in the simulation. This never worked properly
in the transient class.

The break points from devices are updated when a time step is accepted.
However, the stop time, which determines the next discontinuous point
and is used in the time integrator to determine the next step, is not
updated until after the time integrator determines the next step. So
the time integrator does not have the correct stop time information if
a device dynamically inserts a break point. In some cases, the dynamically
added break point is ignored. In other cases, it led to wrong results.
This commit fixes the problem. The fix may cause more update stop
time calls. This may need to be optimized in the transient analysis class