Add a warning message when the stop time is not correct
The transient analysis may not update information correctly if a
device adds a break point before the recent fix. In some cases,
the current time already exceeds stop time which confused the
time integrator and led to wrong behavior. This commit adds a
warning message if the stop time is not correctly updated