Fixes overloading of operator= for block vector classes
The overloaded operator= methods were not correct, so they were not
being used in the block analysis types.  This will be an issue when
the underlying Epetra_MultiVector is separated from the base
Linear::MultiVector class.  This has been corrected with this fix.
Also additional measures have been added to the operator= to exit
when the (multi)vector has no length.

Gitlab issue #36