Make EQN Measures Work with FFT-Based Measures
This commit allows EQN/PARAM measures, for both TRAN and FFT
measure modes, to work with FFT-based measures as part of their
output variables.  Because of the timing of when EQN measures
are updated, the calculation of each FFT analysis' results now
occurs, as soon as the stopTime_ is found (or exceeded), in the
function FFTAnalysis::updateFFTData().  As a backup, those
calculations can also occur in FFTAnalysis::outputResults().  As
part of that change, some of the calculation code was moved to
FFTAnalysis::calculateResults_().

The measure operator creations occur before the function call to
Manager::fixupFFTMeasures().  So, the getMeasureResult() functions
in the FFT measure classes now return the default value if that FFT
measure object has not yet been associated with an FFTAnalysis object.

Both TRAN and FFT mode EQN/PARAM measures are treated as TRAN mode
measures.  That condition is enforced in the constructor for the
EquationEvaluation objects.  So, the Manager::fixupFFTMeasures()
function is basically a noop for EQN/PARAM measures, even if they
were entered via .MEASURE FFT line.

The final change was a re-ordering of the FFT vs. measure calculations
in Manager::remeasure().  The FFT results must be calculated first
to make EQN measures work with FFT-based measures.

SON Bugs1280 and 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327