Fix transmission line breakpointing/convergence
Issue #103 fixed a number of "glitches" in the transmission line
resulting from bad interpolation, but in the course of fixing it we
noticed that sometimes breakpoints got lost, and sometimes the time
integration tried to take a time step longer than the time delay of
the device after a discontinuity.

Ting tracked this down to bad breakpoint management and timestep
selection logic in the transient analysis package.  She fixed that
logic and also patched up the nonlinear solver to do the right thing
when a device says it is never converged (previously, if a device
never said it was converged and the nonlinear solver hit its maximum
number of iterations, it performed all its other tests and would
succeed anyway if those other tests passed).

This was compounded by the delay device and transmission line only
checking for discontinuities in their history *after* a time step was
converged and accepted.  When the transmission line was first
written (nearly 20 years ago) it was forcing a maximum time step equal
to the delay, so that was OK then --- but once that maximum stopped
being applied it was wholly inadequate, and one must detect the
discontinuity *before* taking a step longer than TD.

Ting then added an isConverged method to the Delay device that refuses
to say it's converged if it detects that a time step in excess of the
delay has been taken after a previously undetected discontinuity.
This causes the step to be rejected and the time integration to scale
back to a smaller time step.  Eventually it'll scale back to something
less than the delay and things will proceed correctly with the right
breakpoints set.

This commit makes the same sort of change to the Transmission Line,
which is a little more complex but not much.

With this change, the small "glitches" that were showing up in the
BUG_1370_SON test case are gone (they are too small to matter, and were
previously shrugged off as being a sign of a bug elsewhere --- that
bug is now fixed).  These glitches were in fact due to the
breakpointing logic *and* the fact that it was possible to take a step
longer than the delay before noticing there was a discontinuity.

Fixes #143 (https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/143)
Improves #103 (https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/103)
Related to #122 (https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/122)