Add CMake checks for Basker interface and LOCA
Xyce requires the LOCA subpackage under NOX. Trilinos does not have a
flag indicating whether it is there, so we are checking for the `LOCA.H`
header file.

After the release of Trilinos 12.12, the interface to the Basker solver
in Amesos2 was changed. This commit also adds a check for that interface
to enable a preprocessor flag.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/1