Fix MSVC library peculiarities for CMake
This commit fixes two issues associated with the Microsoft Visual C++
compilers, related to the fact that the MSVC compilers have separate
runtime libraries for debug vs. production builds. As a result,
compilers targeting the MSVC ABI have flags to select the MSVC runtime
library. To deal with this, older versions of CMake set the build type
as part of the `project` call. (Versions 3.15 and later can set
intermediate variables that users can apply themselves, as needed.
However, that's not the default behavior. See, e.g.,
<https://cmake.org/cmake/help/v3.20/policy/CMP0091.html>.) The default
build (if CMAKE_BUILD_TYPE is not set) is debug. The first part of this
commit moves the CMAKE_BUILD_TYPE setting in front of the `project`
call, making "Release" the default build on all systems.

The above behavior was revealed by the latest CMake commit, which added
the test for `LOCA.H`. Teuchos links to the MSVC runtime library.  Since
LOCA.H includes at least one Teuchos header, the test has to point to
the same library as Trilinos, which, of course, is not debug. However,
`check_include_file_cxx` compiles with debug by default, so it links to
a different library, causing a link conflict with Teuchos (making the
test fail). The solution was to use the same compiler flags as Trilinos.
However, the exact approach was difficult to find, because the two
obvious solutions *prepend* tho flags, so they get overridden by the
debug flags of the check, itself:
  `set (CMAKE_REQUIRED_FLAGS ${Trilinos_CXX_COMPILER_FLAGS})`
and
  `check_include_file_cxx(LOCA.H LOCA_IN_Trilinos ${Trilinos_CXX_COMPILER_FLAGS})`
I finally set `CMAKE_REQUIRED_DEFINITIONS` to
`${Trilinos_CXX_COMPILER_FLAGS}`, which seems to work, but is non-ideal.
Nevertheless, I believe it will be robust.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/1