Enable model binning by default.
Issue #144.

I added model binning to Xyce a year or two ago, but I made it an
option that was "off" by default.  That was mostly me being cautious, as
I implemented it fairly quickly and was concerned I might break
something.   At this point I'm pretty confident that I did not break
anything and turning it "on" be default doesn't break any tests, at
least in serial.

This commit changes the model binning boolean used by the parser to be
"true" by default.  Now, it is no longer necessary to have this line:

.options parser model_binning=true

in the netlist to run circuits that use model binning.