Fix binning code to only execute if it needs to
This work is related to issue 144.

On Friday, I pushed small code changes that turned on binning by
default.  This mostly worked fine, but it caused the
BUG_806_SON/mos2_data_step_globalparam.cir to break in nightly testing.

It broke because in this test case, the L,W parameters (which are used
by binning) were being set via global parameters.   It is understandable
that binning is not trustworthy when the L,W parameters can be changed
later in the run (which is the case with global params).  But that
should not have broken this test, as this circuit does NOT do any
binning at all.

This commit fixes the problem.  I have rewritten the function
DeviceBlock::extractModelName, which is in the N_IO_DeviceBlock.C file.
This is one of the main functions involved in model binning.

Before my change, if binning is turned on, then the L, W parameters are
obtained, and then the code attempts to find the name of the model on
the netlist device instance line.   It does this even if the model isn't
relying on binning.  With my change, it now will only attempt to get the
L,W parameter values if it absolutely has to.  That is, if the
conventional model name resolution has failed, and the device in
question is a MOSFET, *then* it will attempt to find the model via
binning, and only *then* will L,W values be obtained.

So, this commit fixes the BUG_806_SON/mos2_data_step_globalparam.cir
test so that it will run with the binning algorithm turned on.  With
this change, I believe the whole test suite passes with binning ON.