Changes Linear::Vector and Linear::MultiVector to be abstract interfaces
This is the final push to change the Linear::Vector and Linear::MultiVector
classes to be abstract interfaces.  The inherited interfaces are found in
the Linear::EpetraVector and Linear::EpetraMultiVector classes, respectively.
While advanced work tried to identify any direct usage of the Linear::Vector
and Linear::MultiVector class and migrate them to the Linear::createVector
and Linear::createMultiVector non-member helper functions, some did slip
through.  These were mostly in the HB analysis, while MOR analysis directly
uses Epetra and should be refactored to perform the calculations through an
abstract interface.

Many classes that use the underlying Epetra objects in the Vector and
MultiVector classes needed to be updated to use the EpetraVectorAccess
interface.  This interface mixes in the methods "epetraObj()" and
"epetraOverlapObj()" for the four Epetra classes: EpetraVector,
EpetraMultiVector, EpetraBlockVector, and EpetraBlockMultiVector.
If a method that receives one of the abstract base classes (Linear,
MultiVector, BlockVector, and BlockMultiVector) needs the underlying
Epetra object, they will need to cast the object to an EpetraVectorAccess
object to get it.

This abstraction has enabled simplification of the block vector classes
because they are used post-assembly (after device loading) and do not
have any notion of "overlap."

Gitlab issue #36