Disable .FFT Processing for non-TRAN Analyses
This commit disables the evaluation of .FFT lines for analysis modes
other than ANP_MODE_TRANSIENT.  This prevents the output of useless
<netlistName>.fftX files.  The fftAnalysisEnabled_ "switch" to
disable FFT processing for the non-TRAN modes is set in the new
function FFTMgr::enableFFTAnalysis(). These FFTMgr functions are
then guarded with that flag: fixupFFTParameters(), UpdateFFTData()
fixupFFTParametersForRemeasure().  This is a bit clunky, but it
means that the Tran outputters don't need to know if they're being
called by a .HB analysis.

The existing FFTMgr::isFFTActive() function also uses that new flag,
since FFTMgr::addFFTAnalysis() is called before Xyce knows what the
analysis mode is.  So, the FFTMgr::FFTAnalysisList still gets made
for all analysis modes.  It's now just not "fixed up", processed or
output for non-TRAN analyses.

SON Bug 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327