Correct std::abs/fabs usage
Last night all the builds failed because "abs" was used all over the
place.  This was "fixed" by replacing those uses with "fabs" but that
was not correct, either, because in some cases the argument is
complex -- the result is compile failures due to automatic type
conversion not working.  The appropriate, templated "abs" function to use is
"std::abs", defined in <complex>.

This commit replaces the few places where fabs() was being called on a
complex number with std::abs().  It actually builds now.