Improve -remeasure for Only .FFT Lines in Netlist
This commit improves the behavior of -remeasure when there are
only .FFT lines in the netlist, without any .MEASURE TRAN or
.MEASURE FFT lines. The -remeasure command line option should work
for that case, but it should not produce an empty <netlistName>.mtX
file for that case.

This commit also makes the usage of isMeasureActive() more uniform
in the measure manager class.  This should be helpful, if I
implement a way for Xyce to (for example) ignore AC measures
during a .TRAN analysis.  The final change is a small change to
the "header text" for the -remeasure output to stdout.  It now
mentions that both .MEASURE and .FFT lines may be re-calculated.

SON Bugs 1280 and 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327