Fix transient analysis to properly update time information and flag
The transient analysis did not update time related information properly. It cannot
handle dynamically inserted break points previously. The fix to the
problem changed the order of things that are updated. It also changed
the beginningIntegration flag when update stop time function is called.
The update stop time not only determines the next discontinous point, but also
updates certain maximum step information. This leads to the different time
stepping in one test which uses non-LTE based error control and compares the verbose
output. The beginningIntegration flag depends on the stop time and the stop time also
depends on this flag. This commit fixes the issue.