Remove RCP dependence from solver state class
Issue #76.

This issue is entitled, "Work out way to get ADMS plugins built without
needing Trilinos headers"

A lot of work on this has happened recently, mostly centering around the
linear algebra objects.  But that work exposed other issues unrelated
to the linear algebra.  One of those issues, that I found out about
yesterday was that my use of an RCP in the
Xyce::Device::SolverState class was also an issue.  That header is
used to build ADMS devices, so having the RCP (which is from Teuchos,
part of Trilinos) was a blocker for this issue.

The RCP in the solver state was being used to point to an expression
group object.  Some devices use expressions and some do not; it was
easiest to have a common group, so I put that group into the solver
state.  Due to several aspects of  new expresison library design,
the expresssion group classes are handled with ref-counted-pointers
(RCP).

This commit fixes the issue.  It fixes it by putting the RCP of the
expressionGroup into a new class called the
Xyce::Device::expressionGroupWrapper.  The wrapper owns the group RCP
now, and the solver state owns a raw pointer to the wrapper.  That way,
the expressionGroupWrapper can just be a foward declaration in the
solver state header.

As there is now an extra layer of indirection, I had to modify all the
device files that use it directly, going from :

getSolverState().expressionGroup_

to:

getSolverState().getGroupWrapper()->expressionGroup_

This commit also adds an include <Teuchos_RCP.hpp> to the
N_DEV_DeviceMgr.h header.  I don't believe that will cause problems for
either the Xyce library mode or for ADMS, however.