Fixes logic error in parallel diagonal block partitioning
A bug was found in the assignment of the diagonal blocks of the
HB block Jacobi preconditioner.  The preconditioner implicitly
assumed there would be at least one diagonal block assigned to
each processor, which is not true if there are more processors
than diagonal blocks (frequencies).  This resulted in an
assignment of zero diagonal blocks per processor and then a
divide by zero.  Once the circuit was run on more processors
than frequencies, the bug exhibited itself.

Gitlab issue #138