Add Initial Text for .MEASURE FFT to Xyce Guides
This commit adds some initial text for the .MEASURE FFT capability
to the Xyce Reference Guide.  That was made into in a separate
section, from .MEASURE, because the qualifers are completely
different.

This commit also updates the text on the -remeasure command line
option in the Reference Guide and Users' Guide.  -remeasure now
works for both .MEASURE and .FFT statements.

SON Bugs 1280 and 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327