Optimize time related information updates in the Transient analysis
The time related information in the transient analysis is used by the
time integrator to determine the next time step. However, some of
the information are updated before the determination of the time step
and some are after. Some updates are in the beginning of transient
loop and some updates are at the end of the loop. This not only can lead
to the wrong information that is communicated to the time integration
because the updates happened at the wrong time (fixed recently), but also
is confusing. This commit changes when the stop time is updated and
it now happens only at the end of the transient loop.

This commit also reduces the number of information updates. Some dependence
between time related function calls requires the information be updated multiple
times. But it only needs to do so in certain cases. The commit fixes
this