Update to allow global params with binning
Related to Issue #144.

Issue #144 is about making binning enabled by default.  This work is
mostly done, but while working on it I discovered that binning with
global parameters was set up (intentially) to never work, and exit
with a fatal error.  It was understandable that I did this.  The use
case I was worried about was this:

You have a netlist in which there are MOSFET instances that have L,W
parameters (used by binning) that are dependent on global params.  For
example:

.global_param width1=3.6u
.global_param width2=1.8u
.global_param length1=1.2u
.global_param length2=3.6u

MLT1 9 Lnot 0 0 NMOS      w={width2} l={length1}

But the netlist is also using binning to determine which MOSFET .MODEL
to use, eg:

.model nmos.1 nmos (level=9 lmin=0.1u lmax=3u  wmin=0.1u wmax=3u )
.model nmos.2 nmos (level=9 lmin=0.1u lmax=3u  wmin=3u  wmax=10u )
.model nmos.3 nmos (level=9 lmin=3u   lmax=10u wmin=0.1u wmax=3u )
.model nmos.4 nmos (level=9 lmin=3u   lmax=10u wmin=3u  wmax=10u )

The W,L instance params are used in binning to find which of the 4
models to select for the instance MLT1.   But, since L,W are set via
global_parameters, they are non-constant and could be updated later.
And, theoretically, they could be changed so much that the original
binning set up during parsing is wrong.

But, I think this scenario case should be made to work for various
reasons.  The first step is making it so that the parser doesn't
barf/exit if L,W are dependent on non-constant expressions.  This
commit accomplishes that step.   Now, the above scenario will
run in Xyce without complaints.

There are 3 use cases (that I can think of) in which this scenario
could happen, two of which don't concern me, and one which does:

1.  global_params set L,W, but the netlist doesn't update L,W
    at all.  In other words, they may as well be constant (and could be
    set via .param, which was already fine before this commit).

2.  global_params set L,W, and the netlist does update L,W
    via .STEP, or .SAMPLING, etc.  But the changes are minor enough that
    they never go outside their original bin.  If using sampling, and
    the variance on L,W is relatively small (relative to the bin size) this
    could plausibly happen.

3.  global_params set L,W, and the netlist does update L,W
    via .STEP, or .SAMPLING, etc, and the changes are big, meaning that
    the bins selected in the parser at some point aren't valid anymore.
    For example, if the bin size is small, and the variance used by
    sampling it big, this could happen.

Before this commit, all of the above use cases would barf in the parser.
Now, all of them will run without complaint.  Cases 1 and 2 will be
fine and I see no obvious reason they would produce incorrect answers.

Case 3 will produce a (slightly?) wrong answer.  Making case 3
be 100% correct will require modifications to the device package that
haven't happened yet, to allow a MOSFET to switch models.
So with my new changes, this is a risk.

However, I've decided that for now that the best choice is to err on the
side of permissiveness.   Not very many people are using the UQ features
of Xyce yet.   And, we are in a transition period in which I am working
towards having .param and .global_param mean the same thing (i.e. that
they are both non-constant).  Finally, in converted netlists (from other
simulators) it is often necessary to make parameters/expressions be
.global_params, because otherwise they violate various Xyce .param rules.
So, I think the above is a use case we need to allow, and then work
towards fixing use case #3.

Fixing this required changing the CircuitContext::fullyResolveParam
function.  This function is ONLY called for binning.  It isn't  used for
anything else.  It is called from the N_IO_DeviceBlock class.

The change involved a call to the param function "getImmutableValue".
If you call that function, and the parameter is dependent on a
non-constant expression, the code exits with a fatal error.  So, I've
changed this section of code to work around this.  Currently, the
N_UTL_Param class does NOT have an equivalent function for "mutable"
values. Such a function may become necessary as I merge the .param and
.global_param capabilities to make them synonymous.